refactor: 소셜 관련 entity 구조에 맞게 이동
