+++ src/main/java/im/toduck/domain/person/SharedRoutine.java
@@ -1,6 +1,6 @@
 package im.toduck.domain.person;
 
-import im.toduck.domain.social.Social;
+import im.toduck.domain.social.persistence.entity.Social;
 import im.toduck.domain.user.persistence.entity.User;
 import im.toduck.global.base.entity.BaseEntity;
 import jakarta.persistence.Entity;
