+++ src/main/java/im/toduck/domain/social/persistence/entity/Comment.java
@@ -1,9 +1,9 @@
-package im.toduck.domain.social;
+package im.toduck.domain.social.persistence.entity;
 
+import im.toduck.domain.user.persistence.entity.User;
 import im.toduck.global.base.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -13,17 +13,21 @@ import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "social_image_file")
+@Table(name = "comment")
 @NoArgsConstructor
-public class SocialImageFile extends BaseEntity {
+public class Comment extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	@Column(nullable = false, length = 1024)
-	private String url;
-	@ManyToOne(fetch = FetchType.LAZY)
+
+	@Column(nullable = false, columnDefinition = "TEXT")
+	private String content;
+	@ManyToOne
+	@JoinColumn(name = "user_id", nullable = false)
+	private User user;
+
+	@ManyToOne
 	@JoinColumn(name = "social_id", nullable = false)
 	private Social social;
-
 }
