+++ src/main/java/im/toduck/domain/social/persistence/entity/SocialCategory.java
@@ -1,4 +1,4 @@
-package im.toduck.domain.social;
+package im.toduck.domain.social.persistence.entity;
 
 import im.toduck.global.base.entity.BaseEntity;
 import jakarta.persistence.Column;
