+++ src/main/java/im/toduck/domain/social/persistence/entity/SocialCategoryLink.java
@@ -1,6 +1,5 @@
-package im.toduck.domain.social;
+package im.toduck.domain.social.persistence.entity;
 
-import im.toduck.domain.user.persistence.entity.User;
 import im.toduck.global.base.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -13,19 +12,19 @@ import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "likes")
+@Table(name = "social_category_link")
 @NoArgsConstructor
-public class Like extends BaseEntity {
+public class SocialCategoryLink extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id", nullable = false)
-	private User user;
-
-	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "social_id", nullable = false)
 	private Social social;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "social_category_id", nullable = false)
+	private SocialCategory socialCategory;
 }
