+++ src/main/java/im/toduck/domain/social/persistence/entity/SocialImageFile.java
@@ -1,6 +1,7 @@
-package im.toduck.domain.social;
+package im.toduck.domain.social.persistence.entity;
 
 import im.toduck.global.base.entity.BaseEntity;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -12,19 +13,17 @@ import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "social_category_link")
+@Table(name = "social_image_file")
 @NoArgsConstructor
-public class SocialCategoryLink extends BaseEntity {
+public class SocialImageFile extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-
+	@Column(nullable = false, length = 1024)
+	private String url;
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "social_id", nullable = false)
 	private Social social;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "social_category_id", nullable = false)
-	private SocialCategory socialCategory;
 }
