+++ src/main/java/im/toduck/domain/social/persistence/entity/SocialTag.java
@@ -1,4 +1,4 @@
-package im.toduck.domain.social;
+package im.toduck.domain.social.persistence.entity;
 
 public enum SocialTag {
 	COMMUNICATION,
